move enum into utils folder
